+++ src/main/java/com/foolsix/fancyenchantments/enchantment/UnyieldingSpirit.java
@@ -3,6 +3,7 @@ package com.foolsix.fancyenchantments.enchantment;
 import com.foolsix.fancyenchantments.FancyEnchantments;
 import com.foolsix.fancyenchantments.capability.TimeToLiveHelper;
 import com.foolsix.fancyenchantments.enchantment.EssentiaEnch.HolyEnchantment;
+import com.foolsix.fancyenchantments.enchantment.handler.LivingDeathEventHandler;
 import com.foolsix.fancyenchantments.enchantment.util.EnchUtils;
 import com.foolsix.fancyenchantments.util.ModConfig;
 import net.minecraft.core.particles.ParticleTypes;
@@ -26,7 +27,7 @@ import net.minecraftforge.event.entity.living.LivingEquipmentChangeEvent;
 import java.util.List;
 
 
-public class UnyieldingSpirit extends HolyEnchantment {
+public class UnyieldingSpirit extends HolyEnchantment implements LivingDeathEventHandler {
     private static final ModConfig.UnyieldingSpiritOptions CONFIG = FancyEnchantments.getConfig().unyieldingSpiritOptions;
     private static final DamageSource GIVING_UP = new DamageSource("fancyenchantments.give_up").bypassArmor();
 
@@ -58,6 +59,11 @@ public class UnyieldingSpirit extends HolyEnchantment {
                 }
             }
         }
+    }
+
+    @Override
+    public void handleLivingDeathEvent(LivingDeathEvent e) {
+        clearTag(e);
     }
 
     public void clearTag(LivingDeathEvent e) {